Front-End/requirements.txt
numpy
pandas
requests
streamlit
altair
streamlit_option_menu
pillow
